feat: change the response format to CustomResponseDto
